Fix and Add
Fixed RSS daily update process.

RSS is now able to have NO daily update if you set RSS update frequency to 0.

Adjust search rules in file library and diary.

Considering the portability of the archived links，I rewrited the process of Link archiving. Now it doesn't generate a url file under the directory of the day, instead it saves the url link in the file_name space (it's only stored in the file_data's keys not in the local file system).